requirements.txt
fastapi==0.115.4
groq==0.11.0
llama-parse==0.5.13
nest-asyncio==1.6.0
pydantic==2.9.2
python-dotenv==1.0.0
uvicorn==0.22.0
